requirements.txt
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pdb-tools==2.5.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
rcsbsearchapi==1.5.1
requests==2.31.0
shiboken6==6.6.2
six==1.16.0
tqdm==4.66.2
urllib3==2.2.1
